+++ docs/release_notes.rst
@@ -19,9 +19,14 @@ The UTCDateTimeAttribute now strictly requires the date string format '%Y-%m-%dT
 PynamoDB has always written values with this format but previously would accept reading other formats.
 Items written using other formats must be rewritten before upgrading.
 
+** Model Serialization **
+
+The ``Model._serialize`` method has changed and now only returns a dictionary of the DynamoDB attribute values.
+
 Other changes in this release:
 
 * Python 2 is no longer supported. Python 3.6 or greater is now required.
+* Table backup functionality (``Model.dump[s]`` and ``Model.load[s]``) has been removed.
 * ``Model.query`` no longer demotes invalid range key conditions to be filter conditions to avoid surprising behaviors:
   where what's intended to be a cheap and fast condition ends up being expensive and slow. Since filter conditions
   cannot contain range keys, this had limited utility to begin with, and would sometimes cause confusing